+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### `0.3.3`
+
+- Fix warnings in dbt-core 1.12
+- Add materialized CTEs for better performance in Duckdb and Clickhouse
+
 ### `0.3.2`
 
 - Address deprecation of `modules.itertools`